Add pcntl extension (removed from core)
The pcntl module is one of several slated [to be moved out of
core](https://wiki.php.net/rfc/unbunle-unmaintained-extensions-php8) in
php8, so it now needs to be installed/enabled manually.

This fixes the "Error: PHP extension pcntl is required for enforcing
time limits" warning that's triggered by ShellCommandBuilder's tests.