Use helper scripts for xdebug too
Updates the xdebug step similarly to the previous pcov/xhprof steps to
use the docker-php-ext-* scripts for installation. Rather than manually
extracting the archive, we can simply clone the given tag. Since we're
using --depth 1 for all the clones, it won't fetch unecessary history.

Also removes the chmod which doesn't seem to be needed when the dir is
created directly within the container instead of being a volume.